Make copy and reference clear at ValueFactory
ValueFactory .newBinary, .newString, .newArray and .newMap take a
mutable argument. Some of them used them as a reference and some of them
copied them. This pull-request makes it clear that they copy the
argument by default and try to omit the copy if an extra argument is
set to true.

This change also includes:

* newMap(Value... kvs) method which was previously newMap(Value[] kvs)
* MessageUnpacker.unpackValue removes a duplicated copy of new byte[]
  (string and binary) or Value[] (array and map).